Add files via upload
Broke the script out into multiple cells to improve readability. I changed some more variable names and filenames for the sake of simplicity. I tested it using the saved searches from today's date and the Wizard excel workbook seems to be working.

Note: I changed the names of some of the files in the 'In' folder